feat: move SubscriptionBanner to the components folder
