+++ src/components/SubscriptionBanner/index.js
